refactor: move related entities code from plugin
